refactor: move helper functions in appropriate modules
moves `get_value_change_triggers` to `Core.Interface` module
