+++ apps/astarte_data_updater_plant/lib/astarte_data_updater_plant/data_updater/core/data_handler.ex
@@ -20,7 +20,6 @@
 
 defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
   alias Astarte.Core.Mapping.ValueType
-  alias Astarte.DataUpdaterPlant.DataUpdater.Impl
   alias Astarte.DataUpdaterPlant.DataUpdater.State
   alias Astarte.DataUpdaterPlant.DataUpdater.CachedPath
   alias Astarte.DataUpdaterPlant.DataUpdater.Cache
@@ -65,7 +64,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
           do: value_timestamp,
           else: div(timestamp, 10000)
 
-      Impl.execute_incoming_data_triggers(
+      Core.DataTrigger.execute_incoming_data_triggers(
         state,
         device_id_string,
         interface_descriptor.name,
@@ -78,7 +77,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
       )
 
       {has_change_triggers, change_triggers} =
-        Impl.get_value_change_triggers(state, interface_id, endpoint_id, path, value)
+        Core.Interface.get_value_change_triggers(state, interface_id, endpoint_id, path, value)
 
       previous_value =
         with {:has_change_triggers, :ok} <- {:has_change_triggers, has_change_triggers},
@@ -101,7 +100,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
 
       if has_change_triggers == :ok do
         :ok =
-          Impl.execute_pre_change_triggers(
+          Core.Trigger.execute_pre_change_triggers(
             change_triggers,
             state.realm,
             device_id_string,
@@ -182,7 +181,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
             "base64_payload" => base64_payload
           }
 
-          Impl.execute_device_error_triggers(
+          Core.Trigger.execute_device_error_triggers(
             state,
             "unset_on_datastream",
             error_metadata,
@@ -231,7 +230,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
             "base64_payload" => base64_payload
           }
 
-          Impl.execute_device_error_triggers(
+          Core.Trigger.execute_device_error_triggers(
             state,
             "unset_not_allowed",
             error_metadata,
@@ -241,7 +240,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
         :ok ->
           if has_change_triggers == :ok do
             :ok =
-              Impl.execute_post_change_triggers(
+              Core.Trigger.execute_post_change_triggers(
                 change_triggers,
                 state.realm,
                 device_id_string,
@@ -296,7 +295,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
           "base64_payload" => base64_payload
         }
 
-        Impl.execute_device_error_triggers(
+        Core.Trigger.execute_device_error_triggers(
           state,
           "write_on_server_owned_interface",
           error_metadata,
@@ -327,7 +326,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
           "base64_payload" => base64_payload
         }
 
-        Impl.execute_device_error_triggers(
+        Core.Trigger.execute_device_error_triggers(
           state,
           "invalid_interface",
           error_metadata,
@@ -356,7 +355,12 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
           "base64_payload" => base64_payload
         }
 
-        Impl.execute_device_error_triggers(state, "invalid_path", error_metadata, timestamp)
+        Core.Trigger.execute_device_error_triggers(
+          state,
+          "invalid_path",
+          error_metadata,
+          timestamp
+        )
 
         update_stats(state, interface, nil, path, payload)
 
@@ -382,7 +386,12 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
           "base64_payload" => base64_payload
         }
 
-        Impl.execute_device_error_triggers(state, "mapping_not_found", error_metadata, timestamp)
+        Core.Trigger.execute_device_error_triggers(
+          state,
+          "mapping_not_found",
+          error_metadata,
+          timestamp
+        )
 
         update_stats(state, interface, nil, path, payload)
 
@@ -407,7 +416,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
           "base64_payload" => base64_payload
         }
 
-        Impl.execute_device_error_triggers(
+        Core.Trigger.execute_device_error_triggers(
           state,
           "interface_loading_failed",
           error_metadata,
@@ -438,7 +447,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
           "base64_payload" => base64_payload
         }
 
-        Impl.execute_device_error_triggers(
+        Core.Trigger.execute_device_error_triggers(
           state,
           "ambiguous_path",
           error_metadata,
@@ -470,7 +479,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
           "base64_payload" => base64_payload
         }
 
-        Impl.execute_device_error_triggers(
+        Core.Trigger.execute_device_error_triggers(
           state,
           "undecodable_bson_payload",
           error_metadata,
@@ -502,7 +511,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
           "base64_payload" => base64_payload
         }
 
-        Impl.execute_device_error_triggers(
+        Core.Trigger.execute_device_error_triggers(
           state,
           "unexpected_value_type",
           error_metadata,
@@ -534,7 +543,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
           "base64_payload" => base64_payload
         }
 
-        Impl.execute_device_error_triggers(
+        Core.Trigger.execute_device_error_triggers(
           state,
           "value_size_exceeded",
           error_metadata,
@@ -566,7 +575,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
           "base64_payload" => base64_payload
         }
 
-        Impl.execute_device_error_triggers(
+        Core.Trigger.execute_device_error_triggers(
           state,
           "unexpected_object_key",
           error_metadata,
@@ -675,7 +684,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Core.DataHandler do
     end
   end
 
-  defp update_stats(state, interface, major, path, payload) do
+  def update_stats(state, interface, major, path, payload) do
     exchanged_bytes = byte_size(payload) + byte_size(interface) + byte_size(path)
 
     :telemetry.execute(
